requirements.txt
# Core
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# TTS Models
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
gTTS>=2.5.0

# Monitoring
psutil>=5.9.0
GPUtil>=1.4.0

# Data Processing
pandas>=2.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Testing
pytest>=7.3.0
